requirements/dev.txt
-r test.txt

flake8==3.3.0
django-debug-toolbar==1.8
django-extensions==1.8.1

